+++ src/SIL.XForge.Scripture/ClientApp/src/assets/i18n/non_checking_en.json
@@ -55,12 +55,14 @@
     "connect_network_to_connect_project": "Please connect to the internet to enable connecting to a new project",
     "my_projects": "My Projects",
     "connect": "Connect",
+    "connect_to_new_project": "Connect to New Project",
     "connecting_to_project": "Connecting to {{ projectName }}...",
     "enable_community_checking": "Enable Community Checking",
     "enable_translation_suggestions": "Enable translation suggestions",
     "engage_the_wider_community_to_check_scripture": "Engage the wider community to ensure that the translation is clear, accurate, and natural. Select verses, ask targeted questions, allow discussion of answers.",
     "error_fetching_resources": "There was an error fetching the Digital Bible Library resources. You will only be able to select projects.",
     "log_in_with_paratext": "Log in with Paratext",
+    "no_connected_projects": "Looks like you haven't connected to any projects yet.",
     "no_connectable_projects": "Looks like there are no connectable projects for you.{{ newLine }}Please go to Paratext to register a project to connect to.",
     "only_paratext_admins_can_start": "** You must be the Paratext Administrator of a project to start it here. Ask your Administrator to connect to your project then you can also connect to it here after that.",
     "paratext_account_linked_to_another_user": "You have logged into a Paratext account that is already linked to a user with the email {{ email }}. You will be logged out of your current account and proceed as the Paratext user.",
